+++ src/components/Modal/Modal.example.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import { Button } from '../../components';
 import Modal from './Modal';
 
@@ -8,44 +8,49 @@ const onManageDisableScrolling = (componentId, scrollingDisabled = true) => {
   console.log('Toggling Modal - scrollingDisabled currently:', componentId, scrollingDisabled);
 };
 
-class ModalWrapper extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { isOpen: false };
-    this.handleOpen = this.handleOpen.bind(this);
-  }
+const ModalWrapper = props => {
+  const [isOpen, setOpen] = useState(false);
 
-  handleOpen() {
-    this.setState({ isOpen: true });
-  }
+  const handleOpen = () => {
+    setOpen(true);
+  };
 
-  render() {
-    return (
-      <div>
-        <div style={{ margin: '1rem' }}>Wrapper text before ModalInMobile</div>
-        <Modal
-          {...this.props}
-          isOpen={this.state.isOpen}
-          onClose={() => {
-            this.setState({ isOpen: false });
-            console.log('Closing modal');
-          }}
-          onManageDisableScrolling={onManageDisableScrolling}
-        >
-          <div style={{ margin: '1rem' }}>Some content inside Modal component</div>
-        </Modal>
-        <div style={{ margin: '1rem' }}>
-          <Button onClick={this.handleOpen}>Open</Button>
-        </div>
+  return (
+    <div>
+      <div style={{ margin: '1rem' }}>Wrapper text before Modal</div>
+
+      <Modal
+        {...props}
+        isOpen={isOpen}
+        onClose={() => {
+          setOpen(false);
+          console.log('Closing modal');
+        }}
+        onManageDisableScrolling={onManageDisableScrolling}
+      >
+        <div style={{ margin: '1rem' }}>Some content inside Modal component</div>
+      </Modal>
+
+      <div style={{ margin: '1rem' }}>
+        <Button onClick={handleOpen}>Open</Button>
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
-export const Empty = {
+export const OldModal = {
   component: ModalWrapper,
   useDefaultWrapperStyles: false,
   props: {
-    id: 'ExampleModal',
+    id: 'OldModal',
+  },
+};
+
+export const ModalWithPortal = {
+  component: ModalWrapper,
+  useDefaultWrapperStyles: false,
+  props: {
+    id: 'ModalWithPortal',
+    usePortal: true,
   },
 };