+++ src/containers/AuthenticationPage/AuthenticationPage.js
@@ -241,6 +241,7 @@ export class AuthenticationPageComponent extends Component {
               id="AuthenticationPage.tos"
               isOpen={this.state.tosModalOpen}
               onClose={() => this.setState({ tosModalOpen: false })}
+              usePortal
               onManageDisableScrolling={onManageDisableScrolling}
             >
               <div className={css.termsWrapper}>