+++ CHANGELOG.md
@@ -14,12 +14,21 @@ way to update this template, but currently, we follow a pattern:
 
 ## Upcoming version 2020-XX-XX
 
+- [change] Update `Modal` component to have option to use `Portal` with `usePortal` flag. Keep also
+  possibility to use modals without Portal because of `ModalInMobile` component.
+  [#1258](https://github.com/sharetribe/ftw-daily/pull/1258)
+
+## [v4.2.0] 2020-02-18
+
 - [add] Show a banner when a user is logged in with limited access.
   [#1259](https://github.com/sharetribe/ftw-daily/pull/1259)
+  [#1261](https://github.com/sharetribe/ftw-daily/pull/1261)
 - [add] Support for logging in as a user from Console.
   [#1254](https://github.com/sharetribe/ftw-daily/pull/1254)
 - [change] Add `handlebars` 4.5.3 and `serialize-javascript` 2.1.1 to resolutions in `package.json`.
   [#1251](https://github.com/sharetribe/ftw-daily/pull/1251)
+
+  [v4.2.0]: https://github.com/sharetribe/flex-template-web/compare/v4.1.0...v4.2.0
 
 ## [v4.1.0] 2020-02-03
 
@@ -29,7 +38,7 @@ way to update this template, but currently, we follow a pattern:
 - [add] Add missing countries (e.g. MX and JP) to `StripeBankAccountTokenInput` validations.
   [#1250](https://github.com/sharetribe/ftw-daily/pull/1250)
 
-  [v4.0.1]: https://github.com/sharetribe/flex-template-web/compare/v4.0.0...v4.0.1
+  [v4.0.1]: https://github.com/sharetribe/flex-template-web/compare/v4.0.0...v4.1.0
 
 ## [v4.0.0] 2019-12-19